requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.10.4
pydantic-settings==2.7.1
firebase-admin==6.6.0
markdown-it-py[plugins]==3.0.0
mdit-py-plugins==0.4.2
linkify-it-py==2.0.3
Pygments==2.18.0
Pillow==11.1.0
Jinja2==3.1.5
bcrypt==4.2.1
httpx==0.28.1
dnspython==2.7.0
python-multipart==0.0.20
gunicorn==23.0.0
mcp==1.3.0
